+++ .gitignore
@@ -36,4 +36,4 @@ out/
 ### VS Code ###
 .vscode/
 
-.env
+.env.DS_Store
