+++ build.gradle
@@ -35,7 +35,7 @@ dependencies {
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-	implementation 'com.fhsh.daitda:common:0.1.1-SNAPSHOT'
+	implementation 'com.fhsh.daitda:common:0.1.2-SNAPSHOT'
 	compileOnly 'org.projectlombok:lombok'
 	runtimeOnly 'org.postgresql:postgresql'
 	annotationProcessor 'org.projectlombok:lombok'
