+++ src/main/java/com/fhsh/daitda/delivery/domain/enums/DeliveryRouteStatus.java
@@ -1,4 +1,4 @@
-package com.fhsh.daitda.domain.enums;
+package com.fhsh.daitda.delivery.domain.enums;
 
 public enum DeliveryRouteStatus {
     HUB_WAITING, HUB_IN_TRANSIT, HUB_ARRIVED, DELIVERING, COMPLETE
