+++ src/main/java/com/fhsh/daitda/delivery/domain/vo/Address.java
@@ -1,11 +1,10 @@
-package com.fhsh.daitda.domain.vo;
+package com.fhsh.daitda.delivery.domain.vo;
 
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.apache.commons.lang3.builder.HashCodeExclude;
 
 @Embeddable
 @Getter
